2024-03-18 Custom Exception 생성 - id 값 null check

+++ src/main/java/com/spring/restapi/board/controller/BoardController.java
@@ -3,6 +3,7 @@ package com.spring.restapi.board.controller;
 import com.spring.restapi.core.dto.response.SuccessResponse;
 import com.spring.restapi.board.dto.request.BoardRequest;
 import com.spring.restapi.board.service.BoardService;
+import com.spring.restapi.core.exception.EmptyIdValueException;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -23,7 +24,10 @@ public class BoardController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<?> getBoard(@PathVariable("id") Long id ) {
+    public ResponseEntity<?> getBoard(@PathVariable(value = "id", required = false) Long id ) {
+        if(id == null) {
+            throw new EmptyIdValueException("BoardController.getBoard::");
+        }
         return SuccessResponse.entity(service.getBoard(id));
     }
 
@@ -39,6 +43,9 @@ public class BoardController {
 
     @DeleteMapping("/{id}")
     public void deleteBoard(@PathVariable("id") Long id) {
+        if(id == null) {
+            throw new EmptyIdValueException("BoardController.getBoard::");
+        }
         service.deleteBoard(id);
     }
